pyproject.toml
[project]
name = "lanista"
version = "0.1.0"
description = "LLM model catalog and router"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    "polars>=1.0",
]

[project.scripts]
lanista = "lanista.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lanista"]

[tool.hatch.build]
include = ["lanista/data/*.json"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["B008"]

[tool.pytest.ini_options]
addopts = "-v --tb=short"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]
